requirements.txt
# Core AI/LLM Libraries
langchain==0.1.0
langchain-openai==0.0.5
langchain-anthropic==0.0.2
langchain-google-genai==0.0.5
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2

# Web Framework
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
websockets==12.0
httpx==0.26.0

# Task Queue
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Database
motor==3.3.2
pymongo==4.6.1
beanie==1.24.0

# Image/Video Generation Services
pillow==10.2.0
requests==2.31.0
aiohttp==3.9.3

# Environment & Config
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0

# Utilities
pandas==2.2.0
numpy==1.26.4
python-dateutil==2.8.2
pytz==2024.1

# Development
pytest==8.0.0
black==24.1.1
flake8==7.0.0